ml/requirements_url_detector.txt
# Core ML and NLP packages
torch>=1.9.0
transformers>=4.20.0
sentence-transformers>=2.2.0
spacy>=3.4.0
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0

# Web scraping and URL processing
newspaper3k>=0.2.8
beautifulsoup4>=4.10.0
requests>=2.25.0
tldextract>=3.1.0
validators>=0.18.0

# Fix for lxml dependency issue
lxml[html_clean]>=4.9.0
# Alternative if the above doesn't work:
# lxml>=4.9.0
# lxml_html_clean>=1.0.0

# Text processing
textblob>=0.15.3
nltk>=3.6.0

# Model persistence
joblib>=1.1.0

# Additional utilities
python-dotenv>=0.19.0
